Implements trace.js
system-trace is an extension that performs a trace on the registry as
things loads. It provides the api described in #1, among other things
allowing you to get the `load` object for each module, get a module's
dependencies and dependants, and walk up the parent map to get the
top-level bundles.

Fixes #1